@@ -20,8 +20,8 @@
   },
   "homepage": "https://github.com/maplibre/ngx-maplibre-gl#readme",
   "peerDependencies": {
-    "@angular/common": ">= 17.0.0",
-    "@angular/core": ">= 17.0.0",
+    "@angular/common": ">= 17.1.0",
+    "@angular/core": ">= 17.1.0",
     "maplibre-gl": ">= 3.6.0",
     "rxjs": ">= 7.8.1"
   }
